Move ./src/stories -> ./stories
Storybook sources are meant for development. By moving them out of
./src, we 1) avoid building them as part of the core package, and
2) avoid publishing them to NPM, which helps cut down our package size.
